+++ src/com/floreysoft/jmte/template/ErrorReportingOutputAppender.java
@@ -5,17 +5,7 @@ import com.floreysoft.jmte.token.Token;
 
 public class ErrorReportingOutputAppender implements OutputAppender {
 
-    private final static String DEFAULT_ERROR_PATTERN = "[!!%s|%s!!]";
-    private final String errorPattern;
-
-    public ErrorReportingOutputAppender() {
-        this(DEFAULT_ERROR_PATTERN);
-    }
-
-    public ErrorReportingOutputAppender(String errorPattern) {
-        this.errorPattern = errorPattern;
-    }
-
+    private final static String ERROR_PATTERN = "[!!%s|%s|%s!!]";
     @Override
     public void append(StringBuilder builder, String text, Token token) {
         final String textToAppend;
@@ -23,7 +13,7 @@ public class ErrorReportingOutputAppender implements OutputAppender {
         if (annotation instanceof JournalingErrorHandler.Entry) {
             final JournalingErrorHandler.Entry entry = (JournalingErrorHandler.Entry) annotation;
             final String message = entry.formattedMessage.formatPlain();
-            textToAppend = String.format(this.errorPattern, message, text);
+            textToAppend = String.format(ERROR_PATTERN, entry.messageKey, message, text);
         } else {
             textToAppend = text;
         }