+++ test/com/floreysoft/jmte/AllTests.java
@@ -5,7 +5,7 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({ InterpretedEngineTest.class, CachedEngineTest.class, CompiledEngineTest.class, MiniParserTest.class,
+@SuiteClasses({ BooleanIfRendererTest.class, ErrorReportingTest.class, InterpretedEngineTest.class, CachedEngineTest.class, CompiledEngineTest.class, MiniParserTest.class,
 		HelperTest.class })
 public class AllTests {